+++ engine/calculus/RaytraceableFunction1D.cs
@@ -15,18 +15,16 @@
  */
 
 using System.Collections.Generic;
-using System.Numerics;
-using System;
 
 namespace FreedomOfFormFoundation.AnatomyEngine.Calculus
 {
-	public abstract class RaytraceableFunction1D : ContinuousMap<float, float>
+	public abstract class RaytraceableFunction1D : ContinuousMap<double, double>
 	{
 		/// <summary>
 		/// Solves the equation \f$(q(x))^2 = b_0 + b_1 x + b_2 x^2 + b_3 x^3 + b_4 x^4\f$, returning all values of
 		///	\f$x\f$ for which the equation is true. \f$q(x)\f$ is the continuous map. The parameters z0 and c
 		///	can be used to substitute x, such that \f$x = z0 + c t\f$. This is useful for raytracing.
 		/// </summary>
-		public abstract IEnumerable<float> SolveRaytrace(QuarticFunction surfaceFunction, float z0 = 0.0f, float c = 1.0f);
+		public abstract IEnumerable<double> SolveRaytrace(QuarticFunction surfaceFunction, double z0 = 0.0, double c = 1.0);
 	}
 }
